# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "BillingOut": {
        "properties": {
          "assistant_paused_by_billing": {
            "title": "Assistant Paused By Billing",
            "type": "boolean"
          },
          "cancel_at_period_end": {
            "title": "Cancel At Period End",
            "type": "boolean"
          },
          "current_period_end": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Period End"
          },
          "current_period_start": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Period Start"
          },
          "interval": {
            "title": "Interval",
            "type": "string"
          },
          "limits": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ]
            },
            "title": "Limits",
            "type": "object"
          },
          "open_invoice": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/OpenInvoice"
              },
              {
                "type": "null"
              }
            ]
          },
          "payment_mode": {
            "title": "Payment Mode",
            "type": "string"
          },
          "pending_interval": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pending Interval"
          },
          "pending_plan": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pending Plan"
          },
          "plan": {
            "title": "Plan",
            "type": "string"
          },
          "plan_name": {
            "title": "Plan Name",
            "type": "string"
          },
          "state": {
            "title": "State",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "trial_days_left": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trial Days Left"
          },
          "trial_ends_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trial Ends At"
          },
          "usage": {
            "$ref": "#/components/schemas/Usage"
          }
        },
        "required": [
          "status",
          "state",
          "plan",
          "plan_name",
          "interval",
          "trial_ends_at",
          "trial_days_left",
          "current_period_start",
          "current_period_end",
          "cancel_at_period_end",
          "pending_plan",
          "pending_interval",
          "usage",
          "limits",
          "open_invoice",
          "assistant_paused_by_billing",
          "payment_mode"
        ],
        "title": "BillingOut",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "InvoiceLine": {
        "properties": {
          "amount_paise": {
            "title": "Amount Paise",
            "type": "integer"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "quantity": {
            "title": "Quantity",
            "type": "integer"
          },
          "unit_paise": {
            "title": "Unit Paise",
            "type": "integer"
          }
        },
        "required": [
          "description",
          "quantity",
          "unit_paise",
          "amount_paise"
        ],
        "title": "InvoiceLine",
        "type": "object"
      },
      "InvoiceOut": {
        "properties": {
          "billing_interval": {
            "title": "Billing Interval",
            "type": "string"
          },
          "currency": {
            "title": "Currency",
            "type": "string"
          },
          "due_at": {
            "format": "date-time",
            "title": "Due At",
            "type": "string"
          },
          "gst_paise": {
            "title": "Gst Paise",
            "type": "integer"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "issued_at": {
            "format": "date-time",
            "title": "Issued At",
            "type": "string"
          },
          "lines": {
            "items": {
              "$ref": "#/components/schemas/InvoiceLine"
            },
            "title": "Lines",
            "type": "array"
          },
          "number": {
            "title": "Number",
            "type": "string"
          },
          "paid_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Paid At"
          },
          "period_end": {
            "format": "date-time",
            "title": "Period End",
            "type": "string"
          },
          "period_start": {
            "format": "date-time",
            "title": "Period Start",
            "type": "string"
          },
          "plan": {
            "title": "Plan",
            "type": "string"
          },
          "status": {
            "enum": [
              "open",
              "paid",
              "void"
            ],
            "title": "Status",
            "type": "string"
          },
          "subtotal_paise": {
            "title": "Subtotal Paise",
            "type": "integer"
          },
          "total_paise": {
            "title": "Total Paise",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "number",
          "status",
          "plan",
          "billing_interval",
          "period_start",
          "period_end",
          "lines",
          "subtotal_paise",
          "gst_paise",
          "total_paise",
          "currency",
          "issued_at",
          "due_at",
          "paid_at"
        ],
        "title": "InvoiceOut",
        "type": "object"
      },
...
      "OpenInvoice": {
        "properties": {
          "due_at": {
            "format": "date-time",
            "title": "Due At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "number": {
            "title": "Number",
            "type": "string"
          },
          "total_paise": {
            "title": "Total Paise",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "number",
          "total_paise",
          "due_at"
        ],
        "title": "OpenInvoice",
        "type": "object"
      },
...
      "PaidInvoice": {
        "properties": {
          "billing_interval": {
            "title": "Billing Interval",
            "type": "string"
          },
          "currency": {
            "title": "Currency",
            "type": "string"
          },
          "due_at": {
            "format": "date-time",
            "title": "Due At",
            "type": "string"
          },
          "gst_paise": {
            "title": "Gst Paise",
            "type": "integer"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "issued_at": {
            "format": "date-time",
            "title": "Issued At",
            "type": "string"
          },
          "lines": {
            "items": {
              "$ref": "#/components/schemas/InvoiceLine"
            },
            "title": "Lines",
            "type": "array"
          },
          "number": {
            "title": "Number",
            "type": "string"
          },
          "paid_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Paid At"
          },
          "payment": {
            "$ref": "#/components/schemas/PaymentOutcome"
          },
          "period_end": {
            "format": "date-time",
            "title": "Period End",
            "type": "string"
          },
          "period_start": {
            "format": "date-time",
            "title": "Period Start",
            "type": "string"
          },
          "plan": {
            "title": "Plan",
            "type": "string"
          },
          "status": {
            "enum": [
              "open",
              "paid",
              "void"
            ],
            "title": "Status",
            "type": "string"
          },
          "subtotal_paise": {
            "title": "Subtotal Paise",
            "type": "integer"
          },
          "total_paise": {
            "title": "Total Paise",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "number",
          "status",
          "plan",
          "billing_interval",
          "period_start",
          "period_end",
          "lines",
          "subtotal_paise",
          "gst_paise",
          "total_paise",
          "currency",
          "issued_at",
          "due_at",
          "paid_at",
          "payment"
        ],
        "title": "PaidInvoice",
        "type": "object"
      },
...
      "PaymentOutcome": {
        "properties": {
          "instructions": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Instructions"
          },
          "mode": {
            "title": "Mode",
            "type": "string"
          },
          "status": {
            "enum": [
              "succeeded",
              "pending",
              "failed"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status",
          "mode"
        ],
        "title": "PaymentOutcome",
        "type": "object"
      },
...
      "SubscribeIn": {
        "additionalProperties": false,
        "properties": {
          "interval": {
            "default": "month",
            "enum": [
              "month",
              "year"
            ],
            "title": "Interval",
            "type": "string"
          },
          "plan": {
            "title": "Plan",
            "type": "string"
          }
        },
        "required": [
          "plan"
        ],
        "title": "SubscribeIn",
        "type": "object"
      },
      "SubscribeOut": {
        "properties": {
          "invoice": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/InvoiceOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "scheduled": {
            "title": "Scheduled",
            "type": "boolean"
          }
        },
        "required": [
          "scheduled",
          "invoice"
        ],
        "title": "SubscribeOut",
        "type": "object"
      },
...
      "Usage": {
        "properties": {
          "conversations": {
            "title": "Conversations",
            "type": "integer"
          },
          "extra_conversations": {
            "title": "Extra Conversations",
            "type": "integer"
          },
          "included_conversations": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Included Conversations"
          },
          "numbers": {
            "title": "Numbers",
            "type": "integer"
          },
          "overage_estimate_paise": {
            "title": "Overage Estimate Paise",
            "type": "integer"
          },
          "period_end": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Period End"
          },
          "period_start": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Period Start"
          },
          "products": {
            "title": "Products",
            "type": "integer"
          },
          "team": {
            "title": "Team",
            "type": "integer"
          }
        },
        "required": [
          "conversations",
          "products",
          "numbers",
          "team",
          "included_conversations",
          "extra_conversations",
          "overage_estimate_paise",
          "period_start",
          "period_end"
        ],
        "title": "Usage",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Versioned REST API. The tenant is always derived from the authenticated session; floor prices are write-only.",
    "title": "WhatsApp AI Sales Assistant API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/billing": {
      "get": {
        "operationId": "billing_api_v1_billing_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BillingOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "The subscription, usage and what is due",
        "tags": [
          "billing"
        ]
      }
    },
    "/api/v1/billing/cancel": {
      "post": {
        "operationId": "cancel_api_v1_billing_cancel_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BillingOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cancel at the end of the paid period (or end the trial now)",
        "tags": [
          "billing"
        ]
      }
    },
    "/api/v1/billing/invoices": {
      "get": {
        "operationId": "invoices_api_v1_billing_invoices_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/InvoiceOut"
                  },
                  "title": "Response Invoices Api V1 Billing Invoices Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Invoices",
        "tags": [
          "billing"
        ]
      }
    },
    "/api/v1/billing/invoices/{invoice_id}/document": {
      "get": {
        "operationId": "document_api_v1_billing_invoices__invoice_id__document_get",
        "parameters": [
          {
            "in": "path",
            "name": "invoice_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Invoice Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "A printable tax invoice",
        "tags": [
          "billing"
        ]
      }
    },
    "/api/v1/billing/invoices/{invoice_id}/pay": {
      "post": {
        "operationId": "pay_api_v1_billing_invoices__invoice_id__pay_post",
        "parameters": [
          {
            "in": "path",
            "name": "invoice_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Invoice Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaidInvoice"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Pay an open invoice",
        "tags": [
          "billing"
        ]
      }
    },
    "/api/v1/billing/resume": {
      "post": {
        "operationId": "resume_api_v1_billing_resume_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BillingOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Keep the subscription after all",
        "tags": [
          "billing"
        ]
      }
    },
    "/api/v1/billing/subscribe": {
      "post": {
        "operationId": "subscribe_api_v1_billing_subscribe_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubscribeIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubscribeOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Choose a plan: upgrades bill now, downgrades apply at the end of the period",
        "tags": [
          "billing"
        ]
      }
    },
...
  }
}